+++ tests/_run.sql
@@ -10,7 +10,15 @@ BEGIN
         -- >>> Run tests
         -- PERFORM fetchq_test.__run('mnt_01', 'It should be possible to run a maintanance job for all the existing queues');
         -- PERFORM fetchq_test.__run('queue_truncate_01', 'It should truncate a queue by name');
-        PERFORM fetchq_test.__run('doc_reschedule_03', 'It should reschedule a document as pending');
+        -- PERFORM fetchq_test.__run('doc_reschedule_03', 'It should reschedule a document as pending');
+        PERFORM fetchq_test.__run('doc_pick_01', '');
+        PERFORM fetchq_test.__run('doc_pick_02', '');
+        PERFORM fetchq_test.__run('doc_pick_03', '');
+        PERFORM fetchq_test.__run('doc_pick_04', '');
+        PERFORM fetchq_test.__run('doc_pick_05', '');
+        PERFORM fetchq_test.__run('doc_pick_06', '');
+        PERFORM fetchq_test.__run('doc_pick_07', '');
+        PERFORM fetchq_test.__run('doc_pick_08', '');
         -- <<<
 
     EXCEPTION WHEN OTHERS THEN
@@ -150,13 +158,13 @@ LANGUAGE plpgsql;
 
 
 -- Define which groups to run
-select * from fetchq_test.__runDevelopement();
+-- select * from fetchq_test.__runDevelopement();
 select * from fetchq_test.__runBasics();
 select * from fetchq_test.__runOptionals();
 
 
 
 -- load tests
--- SELECT * FROM fetchq_test.load_01(100000);
+-- SELECT * FROM fetchq_test.load_01(1000);
 -- SELECT * FROM fetchq_test.load_02(10000);
 -- SELECT * FROM fetchq_test.load_03(25, 10000);